fix(security): address code review findings and improve structure
**Immediate fixes (security/reliability):**
- Fix command injection in android_stage.rs: Add strict validation for host
  (alphanumeric, dots, hyphens only) and launch_component (Android package format)
- Add 10s timeout to ADB commands to prevent hung processes

**Short-term fixes (code quality):**
- Refactor resolume.rs (1,981 lines) into submodule structure:
  - mod.rs: Public API (209 lines)
  - driver.rs: HostDriver implementation (704 lines)
  - clip_map.rs: Clip mapping logic (346 lines)
  - types.rs: Internal types (119 lines)
  - tests.rs: Integration tests (622 lines)
- Document lock acquisition policy in state/mod.rs
- Fix timer race condition: Add new_with_now() and set_target_with_now()
  methods for deterministic testing
