+++ crates/presenter-server/src/resolume/mod.rs
@@ -1,0 +1,209 @@
+mod clip_map;
+mod driver;
+mod types;
+
+use anyhow::anyhow;
+use chrono::{DateTime, Utc};
+use presenter_core::{BibleBroadcast, ResolumeHost, ResolumeHostId};
+use reqwest::Client;
+use serde::Serialize;
+use std::{collections::HashMap, sync::Arc, time::Duration};
+use tokio::sync::{mpsc, RwLock};
+use tokio::task::JoinHandle;
+use tracing::error;
+
+use driver::{run_host_worker, HostCommand};
+
+const DEFAULT_TIMEOUT: Duration = Duration::from_millis(500);
+const HOST_COMMAND_CAPACITY: usize = 16;
+
+#[derive(Debug, Clone, Serialize, PartialEq, Eq)]
+#[serde(rename_all = "snake_case")]
+pub enum ResolumeConnectionState {
+    Disabled,
+    Connecting,
+    Connected,
+    Error,
+}
+
+#[derive(Debug, Clone, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ResolumeConnectionSnapshot {
+    pub state: ResolumeConnectionState,
+    pub last_success: Option<DateTime<Utc>>,
+    pub last_latency_ms: Option<f64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub last_error: Option<String>,
+}
+
+impl ResolumeConnectionSnapshot {
+    pub fn disabled() -> Self {
+        Self {
+            state: ResolumeConnectionState::Disabled,
+            last_success: None,
+            last_latency_ms: None,
+            last_error: None,
+        }
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct StageUpdate {
+    pub current_main: Option<String>,
+    pub current_translation: Option<String>,
+    pub song_name: Option<String>,
+    pub band_name: Option<String>,
+}
+
+#[derive(Debug, Clone)]
+pub struct BibleUpdate {
+    pub passage: Option<BibleBroadcast>,
+}
+
+#[derive(Debug, Clone)]
+pub struct TimerFrame {
+    pub formatted: String,
+}
+
+impl TimerFrame {
+    pub fn new(formatted: String) -> Self {
+        Self { formatted }
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct ResolumeRegistry {
+    client: Client,
+    hosts: Arc<RwLock<HashMap<ResolumeHostId, HostEntry>>>,
+}
+
+#[derive(Debug)]
+struct HostEntry {
+    config: ResolumeHost,
+    status: Arc<RwLock<ResolumeConnectionSnapshot>>,
+    command_tx: mpsc::Sender<HostCommand>,
+    handle: JoinHandle<()>,
+}
+
+impl ResolumeRegistry {
+    pub fn new() -> anyhow::Result<Self> {
+        let client = Client::builder()
+            .timeout(DEFAULT_TIMEOUT)
+            .build()
+            .map_err(|e| anyhow!("failed to build reqwest client: {e}"))?;
+        Ok(Self {
+            client,
+            hosts: Arc::new(RwLock::new(HashMap::new())),
+        })
+    }
+
+    pub async fn set_hosts(&self, hosts: Vec<ResolumeHost>) {
+        let mut guard = self.hosts.write().await;
+        let mut desired: HashMap<ResolumeHostId, ResolumeHost> =
+            hosts.into_iter().map(|host| (host.id, host)).collect();
+
+        // Stop hosts that no longer exist
+        let existing_ids: Vec<_> = guard.keys().copied().collect();
+        for id in existing_ids {
+            if !desired.contains_key(&id) {
+                if let Some(entry) = guard.remove(&id) {
+                    let _ = entry.command_tx.try_send(HostCommand::Shutdown);
+                    entry.handle.abort();
+                }
+            }
+        }
+
+        // Add or update entries
+        for (id, host) in desired.drain() {
+            match guard.get_mut(&id) {
+                Some(entry) => {
+                    if entry.config.host != host.host
+                        || entry.config.port != host.port
+                        || entry.config.is_enabled != host.is_enabled
+                    {
+                        let _ = entry
+                            .command_tx
+                            .try_send(HostCommand::RefreshConfig(host.clone()));
+                        entry.config = host;
+                    } else if entry.config.label != host.label {
+                        entry.config = host;
+                    }
+                }
+                None => {
+                    let entry = self.spawn_host(host);
+                    guard.insert(id, entry);
+                }
+            }
+        }
+    }
+
+    fn spawn_host(&self, host: ResolumeHost) -> HostEntry {
+        let (command_tx, command_rx) = mpsc::channel(HOST_COMMAND_CAPACITY);
+        let status = Arc::new(RwLock::new(if host.is_enabled {
+            ResolumeConnectionSnapshot {
+                state: ResolumeConnectionState::Connecting,
+                last_success: None,
+                last_latency_ms: None,
+                last_error: None,
+            }
+        } else {
+            ResolumeConnectionSnapshot::disabled()
+        }));
+        let client = self.client.clone();
+        let status_clone = Arc::clone(&status);
+        let config_clone = host.clone();
+        let handle = tokio::spawn(async move {
+            if let Err(err) = run_host_worker(client, config_clone, status_clone, command_rx).await
+            {
+                error!(host_id = %host.id, ?err, "resolume worker terminated with error");
+            }
+        });
+        HostEntry {
+            config: host,
+            status,
+            command_tx,
+            handle,
+        }
+    }
+
+    pub async fn stage_update(&self, update: StageUpdate) {
+        for entry in self.hosts.read().await.values() {
+            let _ = entry
+                .command_tx
+                .try_send(HostCommand::Stage(update.clone()));
+        }
+    }
+
+    pub async fn bible_update(&self, update: BibleUpdate) {
+        for entry in self.hosts.read().await.values() {
+            let _ = entry
+                .command_tx
+                .try_send(HostCommand::Bible(update.clone()));
+        }
+    }
+
+    pub async fn timer_update(&self, frame: TimerFrame) {
+        for entry in self.hosts.read().await.values() {
+            let _ = entry.command_tx.try_send(HostCommand::Timer(frame.clone()));
+        }
+    }
+
+    pub async fn snapshot(&self) -> HashMap<ResolumeHostId, ResolumeConnectionSnapshot> {
+        let mut snapshots = HashMap::new();
+        for (id, entry) in self.hosts.read().await.iter() {
+            snapshots.insert(*id, entry.status.read().await.clone());
+        }
+        snapshots
+    }
+
+    pub async fn snapshot_for(&self, id: ResolumeHostId) -> ResolumeConnectionSnapshot {
+        if let Some(entry) = self.hosts.read().await.get(&id) {
+            entry.status.read().await.clone()
+        } else {
+            ResolumeConnectionSnapshot::disabled()
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests;
